Crate noirc_driver

Modules

abi_gen 🔒

stdlib 🔒

Structs

CompileOptions

Contract 🔒

A ‘contract’ in Noir source code with a given name, functions and events. This is not an AST node, it is just a convenient form to return for CrateDefMap::get_all_contracts.

ContractFunctionMeta 🔒

Specifies a contract function and extra metadata that one can use when processing a contract function.

ContractOutputs 🔒

CrateName

Enums

CompileError

CrateId

Constants

DEBUG_CRATE_NAME 🔒

GIT_COMMIT

GIT_DIRTY

NOIRC_VERSION

NOIR_ARTIFACT_VERSION_STRING

Version string that gets placed in artifacts that Noir builds. This is semver compatible. Note: You can’t directly use the value of a constant produced with env! inside a concat! macro.

STD_CRATE_NAME 🔒

Functions

add_debug_source_to_file_manager 🔒

Adds the source code of the debug crate needed to support instrumentation to track variables values

add_dep

Adds a edge in the crate graph for two crates

add_stdlib_source_to_file_manager 🔒

Adds the source code for the stdlib into the file manager

check_crate

Run the def collection, elaboration and type checking passes.

compile_contract

Run the frontend to check the crate for errors then compile all contracts if there were none

compile_contract_inner 🔒

Compile all of the functions associated with a Noir contract.

compile_main

Run the frontend to check the crate for errors then compile the main function if there were none

compile_no_check

Compile the current crate using main_function as the entrypoint.

compute_function_abi

file_manager_with_stdlib

Helper method to return a file manager instance with the stdlib already added

filter_relevant_files

gen_abi

Arranges a function signature and a generated circuit’s return witnesses into a noirc_abi::Abi.

has_errors 🔒

True if there are (non-warning) errors present and we should halt compilation

link_to_debug_crate

prepare_crate

Adds the file from the file system at Path to the crate graph as a root file

prepare_dependency

read_contract 🔒

Return a Vec of all contract declarations in the source code and the functions they contain

ssa_report_to_custom_diagnostic 🔒

stdlib_nargo_toml_source

Returns the contents of the Nargo.toml file for the standard library as a string.

stdlib_paths_with_source

Type Aliases

CompilationResult

Helper type for connecting a compilation artifact to the errors or warnings which were produced during compilation.

ErrorsAndWarnings

Helper type used to signify where errors or warnings are expected in file diagnostics

Warnings

Helper type used to signify where only warnings are expected in file diagnostics