noirc_driver
Function compile_contract

pub fn compile_contract(
    context: &mut Context<'_, '_>,
    crate_id: CrateId,
    options: &CompileOptions,
) -> CompilationResult<CompiledContract>

Run the frontend to check the crate for errors then compile all contracts if there were none