noirc_driver

Function compile_main

pub fn compile_main(
    context: &mut Context<'_, '_>,
    crate_id: CrateId,
    options: &CompileOptions,
    cached_program: Option<CompiledProgram>,
) -> CompilationResult<CompiledProgram>

Run the frontend to check the crate for errors then compile the main function if there were none

On success this returns the compiled program alongside any warnings that were found. On error this returns the non-empty list of warnings and errors.

See compile_no_check for further information about the use of cached_program.