noirc_driver
Function compile_no_check

pub fn compile_no_check(
    context: &mut Context<'_, '_>,
    options: &CompileOptions,
    main_function: FuncId,
    cached_program: Option<CompiledProgram>,
    force_compile: bool,
) -> Result<CompiledProgram, CompileError>

Compile the current crate using main_function as the entrypoint.

This function assumes check_crate is called beforehand.

If the program is not returned from cache, it is backend-agnostic and must go through a transformation pass before usage in proof generation; if it’s returned from cache these transformations might have already been applied.

The transformations are not covered by the check that decides whether we can use the cached artifact. That comparison is based on on CompiledProgram::hash which is a persisted version of the hash of the input ast::Program, whereas the output circuit::Program contains the final optimized ACIR opcodes, including the transformation done after this compilation.