noirc_driver

Type Alias CompilationResult

pub type CompilationResult<T> = Result<(T, Warnings), ErrorsAndWarnings>;

Helper type for connecting a compilation artifact to the errors or warnings which were produced during compilation.

Aliased Type

enum CompilationResult<T> {
    Ok((T, Vec<CustomDiagnostic>)),
    Err(Vec<CustomDiagnostic>),
}

Variants

Ok((T, Vec<CustomDiagnostic>))

Contains the success value

Err(Vec<CustomDiagnostic>)

Contains the error value